ockf done

+++ include/KeyFrame.h
@@ -307,13 +307,12 @@ public:
     void SetKeyFrameDatabase(KeyFrameDatabase* pKFDB);
 
     bool bImu;
-
     // The following variables are accesed from only 1 thread or never change (no mutex needed).
 public:
     // MSPKEYFRAMES
     long unsigned int mReferencecount;
     // mvpOrderedConnectedKeyFrames`
-    long unsigned int mReferencecount_ockf;
+    long long   mReferencecount_ockf;
     //mConnectedKeyFrameWeights
     long unsigned int mReferencecount_ckfw;
     //mObservations
